research/requirements.txt
adbc_driver_manager==1.8.0
adbc_driver_postgresql==1.8.0
adbc_driver_sqlite==1.8.0
aiodns==3.5.0
AppKit==0.2.8
argcomplete==3.6.2
array_api_strict==2.4.1
astroid==3.3.11
async_timeout==5.0.1
asyncssh_unofficial==0.9.2
awscrt==0.28.1
BeaEnginePython==3.1.0
black==25.9.0
blosc==1.11.3
bokeh==3.8.0
boto3==1.40.40
brotli==1.1.0
brotlicffi==1.1.0.0
cachey==0.2.1
cairocffi==1.7.1
capstone==5.0.6
Carbon==1.1.10
cchardet==2.1.7
cerealizer==0.8.4
cityhash==0.4.9
clr==1.0.3
colorama==0.4.6
ConfigParser==7.2.0
coverage==7.10.7
crick==0.0.8
ctypes_snappy==1.03
cudf==0.6.1.post1
curio==1.6
Cython==3.1.4
cytoolz==1.0.1
dask_ml==2025.1.0
de421==2008.1
defcon==0.12.2
defusedxml==0.7.1
dill==0.4.0
distorm3==3.5.2
docrepr==0.2.0
extension_helpers==1.4.0
fastavro==1.12.0
fastparquet==2024.11.0
fitsio==1.2.8
Foundation==0.1.0a0.dev1
fqdn==1.5.1
freetype_py==2.5.1
gevent==25.9.1
glyphsLib==6.11.6
gmpy==1.17
gmpy2==2.2.1
gobject==0.1.0
greenlet==3.2.4
gunicorn==23.0.0
h2==4.3.0
h5==0.12.0
HTMLParser==0.0.2
httpx==0.28.1
huggingface_hub==0.35.1
hypothesis==6.140.2
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipycytoscape==1.3.3
ipyparallel==9.0.1
isal==1.8.0
isoduration==20.11.0
jax==0.7.2
jnius==1.1.0
JPype1==1.6.0
jsonpointer==3.0.0
jsonschema==4.25.1
jupyter_ai==2.31.6
jupyter_ai_magics==2.31.6
kerchunk==0.2.9
keyring==25.6.0
lxml==6.0.2
lz4==4.4.4
lzmaffi==0.3.0
mask==1.0.0
mimesis==18.0.0
mmh3==5.2.0
mpi4py==4.1.0
mtrand==0.1
munkres==1.1.4
mypy==1.18.2
nbformat==5.10.4
ndonnx==0.16.0
netifaces==0.11.0
networkx==3.5
nose==1.3.7
notebook==7.4.7
numbagg==0.9.3
numcodecs==0.16.3
numexpr==2.13.0
numpydoc==1.9.0
objgraph==3.6.2
odfpy==1.4.1
olefile==0.47
openpyxl==3.1.5
pathlib2==2.3.7.post1
pooch==1.8.2
psyco==1.6
pycares==4.11.0
pycurl==7.45.7
pyczmq==0.0.4
pygit2==1.18.2
PyInstaller==6.16.0
pyodide==0.0.2
PyOpenGL==3.1.10
PyQt4==4.11.4
PyQt5==5.15.11
PyQt6==6.9.1
PySide==1.2.4
Pyside==1.2.4
PySide2==5.15.2.1
PySide6==6.9.2
pytest_astropy_header==0.2.2
pytest_fail_slow==0.6.0
pytest_remotedata==0.4.1
pytest_run_parallel==0.7.0
pytest_timeout==2.4.0
pytest_xdist==3.8.0
python_calamine==0.5.3
pyxlsb==1.0.10
qtpy==2.4.3
Quartz==0.0.1.dev0
railroad==0.5.0
regex==2025.9.18
reportlab==4.4.4
requests_kerberos==0.15.0
Res==0.1.7
rfc3339_validator==0.1.4
rfc3986_validator==0.1.1
rfc3987==1.3.8
rmm==0.0.1
sage==0.0.0
scikit_learn==1.7.2
scikits.talkbox==0.2.5
scipy_doctest==2.0.1
selenium==4.35.0
sets==0.3.2
setuptools==80.9.0
setuptools_scm==9.2.0
shiboken2==5.15.2.1
shiboken6==6.9.2
sip==6.12.0
skimage==0.0
skyfield==1.53
smbprotocol==1.15.0
Sphinx==8.2.3
sphinx_gallery==0.19.0
SQLAlchemy==2.0.43
stackless==0.0.1
starlink==0.0.1
sympy==1.14.0
tblib==3.1.0
testpath==0.6.0
threadpoolctl==3.6.0
tiledb==0.35.0
tinycss2==1.4.0
tqdm==4.67.1
trio==0.31.0
ufoLib2==0.18.1
uharfbuzz==0.51.6
ujson==5.11.0
unicodedata2==16.0.0
unittest2==1.1.0
uri_template==1.3.0
urllib3_secure_extra==0.1.0
uvloop==0.21.0
webcolors==24.11.1
win32security==2.1.0
wmi==1.5.1
xarray==2025.9.0
xattr==1.2.0
xlrd==2.0.2
xlsxwriter==3.2.9
xmlrpclib==1.0.1
yapf==0.43.0
zarr==3.1.3
zopfli==0.2.3.post1
zstandard==0.25.0
